Replace values improperly provided as a given string indicating missing value with NA

Replace values improperly provided as a given string indicating missing value with NA

Usage

clean_improper_NAs(
  x,
  possible_strings = c("N/A", "[Nn]on [Rr]enseigné", "[Mm]anquant", "[Uu]nknown",
    "[Nn]ot provided", "[Mm]issing")
)

Arguments

x

the vector

possible_strings

the strings provided in raw table which could possibly indicate a NA value